requirements.txt
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-health-checking==1.76.0
grpcio-testing==1.76.0
grpcio-tools==1.76.0
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
protobuf==6.33.2
Pygments==2.19.2
pytest==9.0.2
setuptools==80.9.0
typing_extensions==4.15.0
